Updated key values to match new js transpose
This should work? I don't really know how to test it, because I'm kind of new to github (embarrassingly). I bet you can figure it out though.

As a sidenote, a really cool feature would be if it automatically updated the output after a small pause, without someone having to hit the button. Also, there is still a bug where an unclosed [ causes it to crash. You may want to revisit regular expressions as a way of cleaning up your text parsing.

Anywho, hope it works and is useful!